refactor: extract groupFindings/canonicalKey to reporters/utils Refs #198
