+++ src/reporters/markdown.ts
@@ -1,6 +1,8 @@
 import { Severity, Pillar, RiskLevel, PILLAR_WEIGHTS } from '../types/index.js';
 import type { ScanReport, Finding, FindingDataSource } from '../types/index.js';
 import { isErrorFinding } from '../issues.js';
+import { groupFindings, MAX_GROUP_REFS } from './utils.js';
+import type { FindingGroup } from './utils.js';
 
 const PILLAR_LABELS: Record<Pillar, string> = {
   [Pillar.SECURITY]: 'Security',
@@ -107,62 +109,6 @@ export interface MarkdownReportOptions {
   grouped?: boolean;
 }
 
-interface FindingGroup {
-  key: string;
-  representative: Finding;
-  members: Finding[];
-}
-
-/**
- * Normalise a finding message into a canonical group key by stripping
- * occurrence-specific details (context suffix, package version, etc.).
- */
-function canonicalKey(f: Finding): string {
-  const msg = f.message;
-
-  // "Non-inclusive term "X" found in <context>" → "Non-inclusive term "X""
-  const iniMatch = msg.match(/^(Non-inclusive term "[^"]+"|Found diminishing language "[^"]+")/);
-  if (iniMatch) return iniMatch[1];
-
-  // "Loosely pinned dependency "pkg": "^x.y.z" uses ^ prefix in devDependencies/..." → canonical
-  if (/Loosely pinned dependency .* uses \^ prefix in devDependencies/.test(msg)) {
-    return 'Loosely pinned devDependencies (^ prefix)';
-  }
-  if (/Loosely pinned dependency .* uses \^ prefix in/.test(msg)) {
-    return 'Loosely pinned dependency (^ prefix)';
-  }
-  if (/Loosely pinned dependency .* uses ~ prefix in devDependencies/.test(msg)) {
-    return 'Loosely pinned devDependencies (~ prefix)';
-  }
-
-  // "Undefined acronym "X" may confuse newcomers" → "Undefined acronym "X""
-  const acronymMatch = msg.match(/^(Undefined acronym "[^"]+")/);
-  if (acronymMatch) return acronymMatch[1];
-
-  // "Assumed knowledge: "X" command used without ..." → "Assumed knowledge: "X" command"
-  const akMatch = msg.match(/^(Assumed knowledge: "[^"]+" (?:command|operation))/);
-  if (akMatch) return akMatch[1];
-
-  // Fallback: use the full message as key (no grouping unless identical)
-  return msg;
-}
-
-function groupFindings(findings: Finding[]): FindingGroup[] {
-  const map = new Map<string, FindingGroup>();
-
-  for (const f of findings) {
-    const key = canonicalKey(f);
-    const existing = map.get(key);
-    if (existing) {
-      existing.members.push(f);
-    } else {
-      map.set(key, { key, representative: f, members: [f] });
-    }
-  }
-
-  return [...map.values()];
-}
-
 /** Render a grouped finding entry for Warnings/Info sections. */
 function renderGroupedFinding(group: FindingGroup): string[] {
   const { key, representative: rep, members } = group;
@@ -185,14 +131,13 @@ function renderGroupedFinding(group: FindingGroup): string[] {
   if (extras.length > 0) lines.push(`  ${extras.join(' ')}`);
 
   // File:line refs — show first 5, then "+N more"
-  const MAX_REFS = 5;
   const withFile = members.filter((m) => m.file);
-  const refs = withFile.slice(0, MAX_REFS).map((m) => {
+  const refs = withFile.slice(0, MAX_GROUP_REFS).map((m) => {
     const loc = m.line ? `${m.file}:${m.line}` : m.file!;
     const ctx = m.context ? ` (${m.context})` : '';
     return `\`${loc}\`${ctx}`;
   });
-  if (withFile.length > MAX_REFS) refs.push(`_+${withFile.length - MAX_REFS} more_`);
+  if (withFile.length > MAX_GROUP_REFS) refs.push(`_+${withFile.length - MAX_GROUP_REFS} more_`);
   if (refs.length > 0) lines.push(`  ${refs.join(' · ')}`);
 
   return lines;
